requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bcrypt==4.1.3
fastapi==0.111.0
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pymongo==4.7.3
